data_integration_agent/requirements.txt
# Data Integration Agent - Dependencies

# Google ADK (Agent Development Kit)
google-adk>=1.0.0

# Google Cloud BigQuery
google-cloud-bigquery>=3.0.0

# Google Cloud Secret Manager
google-cloud-secret-manager>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

# Google Auth (usually included with google-cloud-bigquery)
google-auth>=2.0.0

# Web server
uvicorn[standard]>=0.32.0
fastapi>=0.115.0

# Session persistence with Cloud SQL PostgreSQL
sqlalchemy>=2.0.0
asyncpg>=0.29.0
cloud-sql-python-connector[asyncpg]>=1.0.0
